+++ utils/demo_utils.h
@@ -16,12 +16,12 @@ static inline void progressbar(unsigned int x, unsigned int n, unsigned int w = 
     return;
 
   float ratio = x / (float)n;
-  int c = ratio * w;
+  unsigned int c = (unsigned int)(ratio * w);
 
   std::cout << std::setw(3) << (int)(ratio * 100) << "% [";
-  for (int x = 0; x < c; x++)
+  for (unsigned int x = 0; x < c; x++)
     std::cout << "=";
-  for (int x = c; x < w; x++)
+  for (unsigned int x = c; x < w; x++)
     std::cout << " ";
   std::cout << "]\r" << std::flush;
 }
